+++ libs/asset_type_ec2.py
@@ -18,6 +18,7 @@ class EC2(AssetType):
     """
     def __init__(self, name: str, private_ip: str, public: bool=False):
         super().__init__('EC2', name, public=public)
+        self.os = 'unknown'
         self.private_ip = private_ip
         self.public_ip = ''
         self.security_groups = dict()
@@ -31,6 +32,7 @@ class EC2(AssetType):
             asset_report = self.report_brief()
         else:
             asset_report = {
+                'OS': self.os,
                 'PrivateIP': self.private_ip
             }
             if self.public: